refactor : Controller validation / Test 통합 및 sendRequest 메서드 추출
